LayoutTests/webaudio/biquadfilternode-basic-expected.txt

Basic tests for BiquadFilterNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS BiquadFilterNode has one iutput.
PASS BiquadFilterNode has one output.
PASS Biquad filter defaults to low-pass filter.
PASS The default value of frequency is 350.
PASS The default value of Q is 1.
PASS The default value of gain is 0.
PASS Biquad filter type 0 is settable.
PASS Biquad filter type 1 is settable.
PASS Biquad filter type 2 is settable.
PASS Biquad filter type 3 is settable.
PASS Biquad filter type 4 is settable.
PASS Biquad filter type 5 is settable.
PASS Biquad filter type 6 is settable.
PASS Biquad filter type 7 is settable.
PASS Illegal filter type correctly throws exception.
PASS successfullyParsed is true

TEST COMPLETE
